third_party/WebKit/LayoutTests/fast/events/touch/gesture/long-press-focuses-frame-expected.txt

PASS document.activeElement == frame1 is true
PASS document.activeElement == frame2 is true
PASS document.activeElement == document.body is true
PASS successfullyParsed is true

TEST COMPLETE
This test checks that long press to select text focuses the pressed frame.
 
